src/servers/warranty/src/api_client.py
"""
External API Client for Warranty Data
======================================
Placeholder for external API integrations in MVP.

This module will handle:
- REST API calls to warranty management systems
- Authentication and token management
- Rate limiting and retry logic
- Response caching
"""

import aiohttp
from typing import Optional


class WarrantyAPIClient:
    """Placeholder for external warranty API integration."""
    
    def __init__(self, api_base_url: str = None, api_key: str = None):
        """
        Initialize API client.
        
        Args:
            api_base_url: Base URL for warranty API
            api_key: API authentication key
        """
        self.api_base_url = api_base_url
        self.api_key = api_key
        self.session: Optional[aiohttp.ClientSession] = None
        # TODO: Initialize HTTP session with retry logic
    
    async def get_warranty_status(self, serial_number: str) -> dict:
        """
        Call external API to get warranty status.
        
        Args:
            serial_number: Product serial number
            
        Returns:
            dict: Warranty status from external system
            
        TODO: Implement actual API call
        Example:
            GET /api/v1/warranty/{serial_number}
            Headers: Authorization: Bearer {api_key}
        """
        raise NotImplementedError("API integration pending")
    
    async def validate_product(self, serial_number: str) -> bool:
        """
        Validate product exists in external system.
        
        Args:
            serial_number: Product serial number
            
        Returns:
            bool: True if product exists
            
        TODO: Implement product validation
        """
        raise NotImplementedError("API integration pending")
    
    async def close(self):
        """Close HTTP session."""
        if self.session:
            await self.session.close()
