pyproject.toml
[project]
name = "warranty-poc"
version = "0.1.0"
description = "Warranty Orchestrator POC using Microsoft Agent Framework with MCP support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Scientialibera" }
]
keywords = ["ai", "agent", "azure-openai", "microsoft", "framework", "warranty", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "azure-identity>=1.15.0",
    "azure-core>=1.24.0",
    "openai>=1.0.0",
    "agent-framework>=1.0.0b251223",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "streamlit>=1.28.0",
    "pyjwt[crypto]>=2.8.0",  # OAuth 2.1 token validation
    "requests>=2.31.0",  # JWKS fetching
    "fastapi>=0.104.0",  # MCP server auth middleware
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ==============================================================================
# MCP SERVER CONFIGURATION (HTTP/SSE)
# ==============================================================================

[tool.mcp.servers]
planner = { port = 8001, module = "src.servers.planner_server" }
warranty = { port = 8002, module = "src.servers.warranty.main" }
actions = { port = 8003, module = "src.servers.actions.main" }

# ==============================================================================
# DEVELOPMENT TOOLS
# ==============================================================================

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
